Move edge and node modelling code to /model folder
