+++ src/index.js
@@ -1,7 +1,7 @@
 import * as d3 from 'd3';
 
-import { getPersonNodes, getEntityNodes, setUnknownNode, getNodes } from './nodes/nodes';
-import { checkInterests, getEdges, getOwnershipEdges } from './edges/edges';
+import { getPersonNodes, getEntityNodes, setUnknownNode, getNodes } from './model/nodes/nodes';
+import { checkInterests, getEdges, getOwnershipEdges } from './model/edges/edges';
 import interestTypesCodelist from './codelists/interestTypes';
 import {
   setupD3,
