+++ src/model/nodes/nodeSVGLabel.js
